o  move the path building functions to lib/activate/names.c
o  Update activate.c and fs.c to use them

o  device names are now of the form <vg>:<lv>

+++ lib/activate/fs.c
@@ -14,59 +14,65 @@
 
 #include "fs.h"
 #include "log.h"
+#include "names.h"
 
 #include <libdevmapper.h>
 
-
-void _build_lv_path(char *buffer, size_t len, struct logical_volume *lv,
-		    const char *lv_name)
-{
-	snprintf(buffer, len, "%s/%s_%s", dm_dir(), lv->vg->name, lv_name);
-}
-
-void _build_vg_path(char *buffer, size_t len, struct volume_group *vg)
-{
-	snprintf(buffer, len, "%s%s", vg->cmd->dev_dir, vg->name);
-}
-
-void _build_link_path(char *buffer, size_t len, struct logical_volume *lv,
-		      const char *lv_name)
-{
-	snprintf(buffer, len, "%s%s/%s", lv->vg->cmd->dev_dir,
-		 lv->vg->name, lv_name);
-}
 
 /*
  * Lazy programmer: I'm just going to always try
  * and create/remove the vg directory, and not say
  * anything if it fails.
  */
-static void _mk_dir(struct volume_group *vg)
+static int _mk_dir(struct volume_group *vg)
 {
 	char vg_path[PATH_MAX];
 
-	_build_vg_path(vg_path, sizeof(vg_path), vg);
+	if (!build_vg_path(vg_path, sizeof(vg_path),
+			   vg->cmd->dev_dir, vg->name)) {
+		log_err("Couldn't create volume group directory.");
+		return 0;
+	}
 
 	log_very_verbose("Creating directory %s", vg_path);
 	mkdir(vg_path, 0555);
+
+	return 1;
 }
 
-static void _rm_dir(struct volume_group *vg)
+static int _rm_dir(struct volume_group *vg)
 {
 	char vg_path[PATH_MAX];
 
-	_build_vg_path(vg_path, sizeof(vg_path), vg);
+	if (!build_vg_path(vg_path, sizeof(vg_path),
+			   vg->cmd->dev_dir, vg->name)) {
+		log_err("Couldn't remove volume group directory.");
+		return 0;
+	}
 
 	log_very_verbose("Removing directory %s", vg_path);
 	rmdir(vg_path);
+
+	return 1;
 }
 
 static int _mk_link(struct logical_volume *lv)
 {
 	char lv_path[PATH_MAX], link_path[PATH_MAX];
 
-	_build_lv_path(lv_path, sizeof(lv_path), lv, lv->name);
-	_build_link_path(link_path, sizeof(link_path), lv, lv->name);
+	if (!build_dm_path(lv_path, sizeof(lv_path), lv->vg->name, lv->name)) {
+		log_err("Couldn't create destination path for "
+			"logical volume link.");
+		return 0;
+	}
+
+	if (!build_lv_link_path(link_path, sizeof(link_path),
+				lv->vg->cmd->dev_dir,
+				lv->vg->name, lv->name)) {
+		log_err("Couldn't create source path for "
+			"logical volume link.");
+		return 0;
+	}
 
 	log_very_verbose("Linking %s to %s", link_path, lv_path);
 	if (symlink(lv_path, link_path) < 0) {
@@ -84,7 +90,12 @@ static int _rm_link(struct logical_volume *lv, const char *lv_name)
 	if (!lv_name)
 		lv_name = lv->name;
 
-	_build_link_path(link_path, sizeof(link_path), lv, lv_name);
+	if (!build_lv_link_path(link_path, sizeof(link_path),
+				lv->vg->cmd->dev_dir,
+				lv->vg->name, lv->name)) {
+		log_err("Couldn't create link path (in order to remove it.");
+		return 0;
+	}
 
 	log_very_verbose("Removing link %s", link_path);
 	if (unlink(link_path) < 0) {
@@ -97,9 +108,8 @@ static int _rm_link(struct logical_volume *lv, const char *lv_name)
 
 int fs_add_lv(struct logical_volume *lv)
 {
-	_mk_dir(lv->vg);
-
-	if (!_mk_link(lv)) {
+	if (!_mk_dir(lv->vg) ||
+	    !_mk_link(lv)) {
 		stack;
 		return 0;
 	}
@@ -109,12 +119,11 @@ int fs_add_lv(struct logical_volume *lv)
 
 int fs_del_lv(struct logical_volume *lv)
 {
-	if (!_rm_link(lv, NULL)) {
+	if (!_rm_link(lv, NULL) ||
+	    !_rm_dir(lv->vg)) {
 		stack;
 		return 0;
 	}
-
-	_rm_dir(lv->vg);
 
 	return 1;
 }
